update to latest branch

+++ updater/main/arma3.ps1
@@ -1,5 +1,8 @@
-#Instance Configuration Training Server
-$configPath = 'https://raw.githubusercontent.com/7Cav/service-level-configs/master/mtreck/squad/training/training.json'
+$ConfigPath = $args[0]
+$AutoStart = $args[1]
+$logs = $args[2]
+$Nuke = $args[3]
+
 $configJson = (New-Object System.Net.WebClient).DownloadString($configPath) | ConvertFrom-Json
 #
 $instanceId = $configJson.server.env.SERVER_ID
@@ -18,7 +21,7 @@ $localModDir = $configJson.server.env.LOCAL_MOD_PATH
 $steamUser = $env:STEAM_USER
 $steamPass = $env:STEAM_PASS
 # Join Paths and other
-$storePath += "training"
+$storePath += "$serverName"
 $installDirWorkshop = "$storePath\$installDirWorkshop"
 $ServiceName = "$instanceId"
 $ServiceName += "Updater"
@@ -29,32 +32,43 @@ $serverScripts = "$installDirArmadirectory\$scripts"
 # Remove existing symbolic links
 #
 Write-Output "Update has started: $(Get-Date) for Service $instanceId - $serverName"
-
-$dirp = Get-Item $installDirArmadirectory\@*
-foreach($item in $dirp)
-{
-$item.Delete()
-Write-Output "Removing Junction Point at $item"
-}
-
 #Stop Firedaemon Service
 net stop $instanceId
 Start-Sleep -s 5
 Write-Output Start-Sleep -s 15
+$dirp = Get-Item $installDirArmadirectory\@*
+foreach ($item in $dirp) {
+    $item.Delete()
+    Write-Output "Removing Junction Point at $item"
+}
+# Nuke it all
+if (-e$Nuke q $True) {
+    Write-Output "!!!!!!!!!!"
+    Write-Output "Warning, -Nuke Parameter was used. all files related to this server instance will be wiped"
+    Write-Output "!!!!!!!!!!"
+    $dirp = Get-Item $installDirArmadirectory\*
+    foreach ($item in $dirp) {
+        Remove-Item -Recurse -Force $item
+        Write-Output "Removing from Installation Directory: $item"
+    }
+    $dirS = Get-Item $installDirWorkshop
+    foreach ($item in $dirS) {
+        Remove-Item -Recurse -Force $item
+        Write-Output "Removing from Store Directory: $item"
+    }
+}
 
 Write-Output "Starting SteamCMD Download and Validate Base Installation"
 #login to steamcmd using env variables
-$argumentListArray = "+login $steamUser $steamPass +force_install_dir "+$storePath+" "
+$argumentListArray = "+login $steamUser $steamPass +force_install_dir " + $storePath + " "
 #download each item in steamcmd using the app id in the mod list array
-foreach($item in $modListJson) 
-{
+foreach ($item in $modListJson) {
     $id = $item.app
     $name = $item.name
     $server = $item.server
-    if ($id -ne "")
-    {
-    $argumentListArray += "+workshop_download_item 107410 $id "
-    Write-Output "Found Item_ID: $id ($name) - Servermod = $server"
+    if ($id -ne "") {
+        $argumentListArray += "+workshop_download_item 107410 $id "
+        Write-Output "Found Item_ID: $id ($name) - Servermod = $server"
     }
 }   
 #update and validate arma 3 base installation
@@ -63,23 +77,20 @@ $argumentListArray += "+force_install_dir $installDirArmadirectory +app_update 2
 Start-Process -FilePath $steamCMDdir -ArgumentList $argumentListArray -NoNewWindow -Wait
 
 #copying keys and mods to dir from workshop
-foreach($item in $modListJson) 
-{
-   $id = $item.app
-   $name = $item.name
-   $path = $localModDir
-   if ($id -ne "")
-   {
-      New-Item -ItemType Junction -Path "$installDirArmadirectory\$name" -Target $installDirWorkshop\$id     
-      copy-item $installDirWorkshop\$id\[kK]*\*.bikey $installDirArmadirectory\keys\ -force -recurse
-      Write-Output "Copy Key $name" 
-   }
-   else
-   {
-      New-Item -ItemType Junction -Path "$installDirArmadirectory\$name" -Target "$path\$name"
-      copy-item $installDirArmadirectory\$name\[kK]*\*.bikey $installDirArmadirectory\keys\ -force -recurse
-      Write-Output "Copy Key $name"
-   }
+foreach ($item in $modListJson) {
+    $id = $item.app
+    $name = $item.name
+    $path = $localModDir
+    if ($id -ne "") {
+        New-Item -ItemType Junction -Path "$installDirArmadirectory\$name" -Target $installDirWorkshop\$id     
+        copy-item $installDirWorkshop\$id\[kK]*\*.bikey $installDirArmadirectory\keys\ -force -recurse
+        Write-Output "Copy Key $name" 
+    }
+    else {
+        New-Item -ItemType Junction -Path "$installDirArmadirectory\$name" -Target "$path\$name"
+        copy-item $installDirArmadirectory\$name\[kK]*\*.bikey $installDirArmadirectory\keys\ -force -recurse
+        Write-Output "Copy Key $name"
+    }
 }
 
 Write-Output "$serverScripts has yet to be assigned in the script"
@@ -90,16 +101,13 @@ Write-Output "$serverScripts has yet to be assigned in the script"
 # Instance Mods
 $modparam = "-mod="
 $modserverparam = "-servermod="
-foreach($item in $modListJson)
-{
+foreach ($item in $modListJson) {
     $name = $item.name
     $server = $item.server
-    if ($server -ne "true")
-    {
+    if ($server -ne "true") {
         $modparam += "$name;"
     }
-    else
-    {
+    else {
         $modserverparam += "$name;"
     }
 }
@@ -119,12 +127,32 @@ Write-Output "Parameters: $compileParams"
 Set-ItemProperty -Path 'Registry::HKEY_LOCAL_MACHINE\System\CurrentControlSet\Control\Session Manager\Environment' -Name $serverName -Value $compileParams
 
 #Removing logs after update
-Remove-Item $configDir\*.rpt 
-Remove-Item $configDir\*.log 
-Write-Output Remove-Item $configDir\*.rpt 
-Write-Output Remove-Item $configDir\*.log 
-
+if ($Logs -ne $True) {
+    Write-Output "-logs not found, removing .rpt and .log files from config directory"
+    Remove-Item $configDir\*.rpt 
+    Remove-Item $configDir\*.log 
+    Write-Output Remove-Item $configDir\*.rpt 
+    Write-Output Remove-Item $configDir\*.log S
+}
+else {
+    Write-Output "-Logs found, keeping old .rpt and .log files"
+}
+# missions folder
+if ($serverName -match "training") {
+    New-Item -ItemType Junction -Path "$installDirArmadirectory\$name" -Target D:\gameservers\arma3\mpmissions
+    Write-Output "[NOTICE] symlink to \mpmissions created"
+}
+else {
+    write-output "[NOTICE] Detected non-training server config, symlink for mpmissions skipped"
+}
 Write-Output "Update has finished: $(Get-Date) for $serverName"
+if ($AutoStart -eq $True) {
+    net start $instanceId
+    Write-Output "Starting $instanceId"
+}
+else {
+    Write-Output "-AutoStart was not set, exiting without starting server instance"
+}
 #Stop myself if service Firedaemon Service
 net stop $ServiceName
 [Environment]::Exit(66)
